movies_admin_panel/requirements.txt
django==4.0.4
uwsgi==2.0.20
django-split-settings==1.2.0
python-dotenv==0.21.0
django-debug-toolbar==3.4.0
psycopg2==2.9.4
django_extensions==3.2.1
elasticsearch==7.13.0
pydantic==1.10.2